refactor: 대시보드 추가 모달컴포넌트로 분리
